requirements.txt
pandas~=1.5.2
pytz~=2022.7
bokeh~=3.0.3
Flask~=2.2.2
bson~=0.5.10
pymongo~=4.3.3
requests~=2.28.1
python-dotenv~=0.21.0
confluent-kafka~=1.9.2
